fastapi_kafka/requirements.txt
fastapi
uvicorn
kafka-python
pandas
pyarrow
fastparquet